requirements.txt
absl-py==0.11.0
aiohttp==3.7.3
alembic==1.4.1
argon2-cffi==20.1.0
ase==3.21.1
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
azure-core==1.10.0
azure-storage-blob==12.7.1
backcall==0.2.0
bleach==3.2.3
cached-property==1.5.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
cloudpickle==1.6.0
cryptography==3.3.1
cycler==0.10.0
databricks-cli==0.14.1
datatable==0.11.1
decorator==4.4.2
defusedxml==0.6.0
docker==4.4.1
entrypoints==0.3
Flask==1.1.2
flatbuffers==1.12
fsspec==0.8.5
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.12
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googledrivedownloader==0.4
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
imbalanced-learn==0.8.0
imblearn==0.0
importlib-metadata==3.4.0
iniconfig==1.1.1
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
isort==5.7.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
llvmlite==0.35.0
Mako==1.1.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.4
metis==0.2a5
mistune==0.8.4
mlflow==1.13.1
msrest==0.6.21
multidict==5.1.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numba==0.52.0
numpy==1.20.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.1
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
pluggy==0.13.1
prometheus-client==0.9.0
prometheus-flask-exporter==0.18.1
prompt-toolkit==3.0.14
protobuf==3.14.0
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.7.4
pynndescent==0.5.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-louvain==0.15
python-mnist==0.7
pytorch-lightning==1.1.6
pytz==2020.5
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.0.2
qtconsole==5.0.2
QtPy==1.9.0
querystring-parser==1.2.4
rdflib==5.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==3.0.5
SQLAlchemy==1.3.22
sqlparse==0.4.1
tabulate==0.8.7
tensorboard==2.5.0
tensorboard-data-server==0.6.0
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==1.15.3
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
texttable==1.6.3
threadpoolctl==2.1.0
tifffile==2021.1.14
toml==0.10.2
torch==1.5.0+cu101
torch-cluster @ https://pytorch-geometric.com/whl/torch-1.5.0/torch_cluster-1.5.7+cu101-cp37-cp37m-linux_x86_64.whl
torch-geometric==1.6.3
torch-scatter @ https://pytorch-geometric.com/whl/torch-1.5.0/torch_scatter-2.0.5+cu101-cp37-cp37m-linux_x86_64.whl
torch-sparse @ https://pytorch-geometric.com/whl/torch-1.5.0/torch_sparse-latest+cu101-cp37-cp37m-linux_x86_64.whl
torch-spline-conv @ https://pytorch-geometric.com/whl/torch-1.5.0/torch_spline_conv-latest+cu101-cp37-cp37m-linux_x86_64.whl
torchvision==0.6.0+cu101
tornado==6.1
tqdm==4.56.0
traitlets==5.0.5
typing-extensions==3.7.4.3
umap-learn==0.5.1
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
xlrd==2.0.1
yarl==1.6.3
zipp==3.4.0
